source_code/web/requirements.txt
asgiref==3.7.2
black==23.7.0
Brotli==1.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
contourpy==1.2.0
coverage==7.3.2
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
Django==4.2.4
django-cors-headers==4.2.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.5
django-nose==1.4.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
fonttools==4.44.3
future==0.18.3
google-api-core==2.12.0
google-api-python-client==2.102.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.60.0
graphtools==1.5.3
html5lib==1.1
httplib2==0.22.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mypy-extensions==1.0.0
mysqlclient==2.2.0
nose==1.3.7
numpy==1.26.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.10.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
PyGSP==0.5.1
PyJWT==2.8.0
pyparsing==3.1.1
pyphen==0.14.0
python-dateutil==2.8.2
python-decouple==3.8
python-magic==0.4.3
pytz==2023.3
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
scprep==1.2.3
six==1.16.0
sqlparse==0.4.4
stripe==7.5.0
tasklogger==1.2.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
weasyprint==60.1
webencodings==0.5.1
wrapt==1.16.0
zopfli==0.2.3
